+++ idf_component_tools/manifest/manifest.py
@@ -118,6 +118,7 @@ class Manifest(object):
                 version_spec=details.get('version') or '*',
                 public=details.get('public'),
                 if_clauses=details.get('rules'),
+                require=details.get('require', None),
             )
             if component.meet_optional_dependencies:
                 manifest._dependencies.append(component)
@@ -138,7 +139,7 @@ class Manifest(object):
         if self._manifest_hash:
             return self.manifest_hash
 
-        serialized = self.serialize()  # type: ignore
+        serialized = self.serialize(serialize_default=False)  # type: ignore
         return hash_object(serialized)
 
 
@@ -150,6 +151,11 @@ class ComponentRequirement(object):
         'source',
         'version_spec',
         'meet_optional_dependencies',
+        {
+            'name': 'require',
+            'default': True,
+            'serialize_default': False
+        },
     ]
 
     def __init__(
@@ -159,13 +165,19 @@ class ComponentRequirement(object):
             version_spec='*',  # type: str
             public=None,  # type: bool | None
             if_clauses=None,  # type: list[IfClause] | None
+            require=None,  # type: str | bool | None
     ):
         # type: (...) -> None
         self._version_spec = version_spec
         self.source = source
         self._name = name
-        self.public = public
+        self.public = None  # type: bool | None
+        if require == 'public' or public:
+            self.public = True
+        elif public is False or require == 'private':
+            self.public = False
         self.if_clauses = if_clauses
+        self.require = True if require in ['private', 'public', None] else False
 
     @property
     def meta(self):