+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add the `repository`, `documentation`, `issues`, and `discussion` fields for the URLs in the root of the manifest
 - Cache request to the API component registry
+- Add `require` field for dependencies. Add possibility to download dependencies without building them.
 - Default revision number change from 1 to 0
 - Warn users when the `override_path` point to a non-component directory